Move class SystemAnnouncementEntry to separate file
